pyproject.toml
[project]
name = "ai-travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[google-genai]>=1.9.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.3.3",
]
